+++ Backtrace/Model/BacktraceStackTrace.cs
@@ -1,11 +1,8 @@
 ﻿using Backtrace.Common;
-using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Reflection;
-using System.Text;
 
 namespace Backtrace.Model
 {
@@ -38,11 +35,7 @@ namespace Backtrace.Model
         private void Initialize()
         {
             bool generateExceptionInformation = _exception != null;
-            //initialize environment stack trace
             var stackTrace = new StackTrace(true);
-            //reverse frame order
-            var frames = stackTrace.GetFrames();
-            SetStacktraceInformation(frames, generateExceptionInformation);
             if (_exception != null)
             {
                 if (CallingAssembly == null)
@@ -52,6 +45,12 @@ namespace Backtrace.Model
                 var exceptionStackTrace = new StackTrace(_exception, true);
                 var exceptionFrames = exceptionStackTrace.GetFrames();
                 SetStacktraceInformation(exceptionFrames, true);
+            }
+            else
+            {
+                //reverse frame order
+                var frames = stackTrace.GetFrames();
+                SetStacktraceInformation(frames, generateExceptionInformation);
             }
             //Library didn't found Calling assembly
             //The reason for this behaviour is because we throw exception from TaskScheduler